+++ Library/src/main/java/dev/testify/internal/helpers/FindRootView.kt
@@ -1,0 +1,9 @@
+package dev.testify.internal.helpers
+
+import android.app.Activity
+import android.view.ViewGroup
+import androidx.annotation.IdRes
+import dev.testify.internal.exception.RootViewNotFoundException
+
+fun Activity.findRootView(@IdRes rootViewId: Int): ViewGroup =
+    this.findViewById(rootViewId) ?: throw RootViewNotFoundException(this, rootViewId)