+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Testify Change Log
 
+## Unreleased
+
+### Library
+
+#### Changed
+
+- `ScreenshotRule.getRootView()` is now an extension function `fun Activity.findRootView(@IdRes rootViewId: Int): ViewGroup`
+- `ScreenshotRule.setCaptureMethod()` is deprecated. Use `var captureMethod: CaptureMethod?` on `TestifyConfiguration` to set the capture method.
+- `ScreenshotRule.setCompareMethod()` is deprecated. Use `var compareMethod: CompareMethod?` on `TestifyConfiguration` to set the compare method.
+- `ScreenshotRule.compareBitmaps()` is now a top-level function.
+- `ScreenshotRule.takeScreenshot()` is now a top-level function.
+
 ## 2.0.0-beta03
 
 ### All Projects